+++ src/requisition-view/requisition-stock-count-date-modal.controller.spec.js
@@ -15,28 +15,23 @@
 
 describe('RequisitionStockCountDateModalController', function() {
 
-    var $controller, $q, $rootScope, messageService, vm, requisition, modalDeferred,
-    alertService, message;
+    var $controller, messageService, vm, requisition, modalDeferred, message;
 
     beforeEach(function() {
         module('requisition-view');
 
         inject(function($injector) {
             $controller = $injector.get('$controller');
-            $q = $injector.get('$q');
-            $rootScope = $injector.get('$rootScope');
             messageService = $injector.get('messageService');
-            alertService = $injector.get('alertService');
         });
 
         message = 'some-message';
         spyOn(messageService, 'get').andReturn(message);
-        spyOn(alertService, 'error');
 
         requisition = {
             datePhysicalStockCountCompleted: "2017-08-11"
         };
-        modalDeferred = $q.defer();
+        modalDeferred = jasmine.createSpyObj('modalDeferred', ['resolve']);
 
         vm = $controller('RequisitionStockCountDateModalController', {
             requisition: requisition,
@@ -62,78 +57,53 @@ describe('RequisitionStockCountDateModalController', function() {
     });
 
     describe('submit', function() {
-        beforeEach(function() {
-            var currentDate = new Date();
-            var date = new Date(currentDate.getUTCFullYear(), currentDate.getUTCMonth(), currentDate.getUTCDate() + 1, 0, 0, 0, 0);
 
-            vm.datePhysicalStockCountCompleted = convertToISOLocalDateString(date);
+        it('should close modal', function() {
+            vm.submit();
+
+            expect(modalDeferred.resolve).toHaveBeenCalled();
         });
 
-        it('should set invalidMessage if date is after UTC current date', function () {
-            vm.submit();
-            $rootScope.$apply();
+        it('should update requisition date', function() {
+            var date = moment().toISOString();
 
-            expect(vm.invalidMessage).toEqual(message);
-        });
-
-        it('should call alertService error with proper parameter if date is after UTC current date', function () {
-            vm.submit();
-            $rootScope.$apply();
-
-            expect(alertService.error).toHaveBeenCalledWith(message);
-        });
-
-        it('should resolve modalDeffered if date is at UTC current date', function () {
-            var currentDate = new Date();
-            var date = new Date(currentDate.getUTCFullYear(), currentDate.getUTCMonth(), currentDate.getUTCDate(), 0, 0, 0, 0);
-
-            vm.datePhysicalStockCountCompleted = convertToISOLocalDateString(date);
-
-            var isDeffered = false;
-            modalDeferred.promise.then(function () {
-                isDeffered = true;
-            });
+            vm.datePhysicalStockCountCompleted = date;
 
             vm.submit();
-            $rootScope.$apply();
 
-            expect(isDeffered).toBe(true);
-        });
-
-        it('should resolve modalDeffered if date is before UTC current date', function () {
-            var currentDate = new Date();
-            var date = new Date(currentDate.getUTCFullYear(), currentDate.getUTCMonth() - 1, currentDate.getUTCDate(), 0, 0, 0, 0);
-
-            vm.datePhysicalStockCountCompleted = convertToISOLocalDateString(date);
-
-            var isDeffered = false;
-            modalDeferred.promise.then(function () {
-                isDeffered = true;
-            });
-
-            vm.submit();
-            $rootScope.$apply();
-
-            expect(isDeffered).toBe(true);
-        });
-
-        it('should resolve modalDeffered if date is in past', function () {
-            vm.datePhysicalStockCountCompleted = '2017-08-10';
-
-            var isDeffered = false;
-            modalDeferred.promise.then(function () {
-                isDeffered = true;
-            });
-
-            vm.submit();
-            $rootScope.$apply();
-
-            expect(isDeffered).toBe(true);
+            expect(requisition.datePhysicalStockCountCompleted).toEqual(date);
         });
     });
 
-    function convertToISOLocalDateString(date) {
-        return date.toISOString().slice(0, 10);
-    }
+    describe('validateDate', function() {
+    
+        it('should return undefined if date is undefined', function() {
+            vm.datePhysicalStockCountCompleted = undefined;
+
+            expect(vm.validateDate()).toBeUndefined();
+        });
+
+        it('should return error if date is in the future', function() {
+            messageService.get.andReturn('You cannot record date for the future');
+
+            vm.datePhysicalStockCountCompleted = moment().add(1, 'days').toISOString();
+
+            expect(vm.validateDate()).toEqual('You cannot record date for the future');
+            expect(messageService.get).toHaveBeenCalledWith('requisitionView.datePhysicalStockCountCompleted.inFuture');
+        });
+
+        it('should return undefined if date is in the past', function() {
+            vm.datePhysicalStockCountCompleted = moment().subtract(1, 'days').toISOString();
+
+            expect(vm.validateDate()).toBeUndefined();
+        });
+
+        it('should return undefined for current day', function() {
+            vm.datePhysicalStockCountCompleted = moment().toISOString();
+
+            expect(vm.validateDate()).toBeUndefined();
+        });
+    
+    });
 
 });